feature: base component
