+++ hw/code/ui/components/base_component.py
@@ -1,0 +1,29 @@
+from typing import List
+
+from selenium.webdriver.remote.webelement import WebElement
+from selenium.webdriver.support.wait import WebDriverWait
+from selenium.webdriver.support import expected_conditions as EC
+
+class BaseComponent(object):
+    def wait(self, timeout=None) -> WebDriverWait:
+        """
+        Get WebDriverWait object with given timeout
+        """
+        if timeout is None:
+            timeout = 5
+        return WebDriverWait(self.driver, timeout=timeout)
+
+    def find(self, locator, timeout=None) -> WebElement:
+        """
+        Find the element by locator. Waits for the element to be loaded with given timeout
+        """
+        return self.wait(timeout).until(EC.presence_of_element_located(locator))
+
+    def find_all(self, locator, timeout=None) -> List[WebElement]:
+        self.wait(timeout).until(EC.presence_of_element_located(locator))
+        return self.driver.find_elements(locator)
+
+    def click(self, locator, timeout=None):
+        self.find(locator, timeout=timeout)
+        elem = self.wait(timeout).until(EC.element_to_be_clickable(locator))
+        elem.click()
